+++ client/src/components/meetingCard/MeetingCard.tsx
@@ -5,52 +5,12 @@ import { faCheck } from "@fortawesome/free-solid-svg-icons";
 import { Button } from "reactstrap";
 import { Appointment, User } from "../../services/api.types";
 import { api } from "../../services/api";
+import { BreakDownTime, CapitalizeFirstLetter } from "../../services/tools";
 
 interface IProps {
     appt: Appointment,
     isTutor: boolean,
-}
-
-const monthMap = new Map([
-    [1, "January"],
-    [2, "February"],
-    [3, "March"],
-    [4, "April"],
-    [5, "May"],
-    [6, "June"],
-    [7, "July"],
-    [8, "August"],
-    [9, "September"],
-    [10, "October"],
-    [11, "November"],
-    [12, "December"],
-])
-
-function BreakDownTime(standard_time: String): Array<String> {
-    let date_time = standard_time.split("T")
-
-    let date_arr = date_time[0].split("-")
-    let month = monthMap.get(Number(date_arr[1])) ? monthMap.get(Number(date_arr[1])) : " ";
-    let date = month + " " + date_arr[2] + ", " + date_arr[0]
-
-    let time_arr = date_time[1].split(":");
-    let meridian = " AM"
-    let hour = Number(time_arr[0])
-    if (hour > 12) {
-        hour -= 12
-        meridian = " PM"
-    }
-    let time = String(hour) + ":" + time_arr[1] + meridian;
-    return [date,time];
-}
-
-function CapitalizeFirstLetter(str:String): String {
-    let string_arr = str.split(" ");
-    let new_arr:string[] = []
-    for (const word of string_arr) {
-        new_arr.push(word.charAt(0).toUpperCase() + word.slice(1))
-    }
-    return new_arr.join(" ");
+    includePrevious: boolean,
 }
 
 export function MeetingCard(props: IProps) {
